music/requirements.txt
asgiref==3.6.0
azure-core==1.26.3
azure-identity==1.12.0
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.1
cryptography==39.0.2
Django==4.1.7
django-storages-azure==1.6.8
djangorestframework==3.14.0
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msgpack==1.0.5
mutagen==1.46.0
pika==1.3.1
portalocker==2.7.0
protobuf==4.22.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
